Update receipt result
Move out Receipt class to separate file.
Translate comments.
Mark some properties as obsolete.
Add new properties.
Fix existing properties to correspond current reality.
